Fix CUDA versions shown in output
The cuda runtime api version was printed to the screen, but mumax3
uses only the cuda driver api and thus not the runtime api.

Now, the mumax3 uname contains the cuda driver api version against
which mumax3 is  compiled . If compiled correctly, this should match
the versions of the used cuda libraries.

The gpu info now also contains the latest cuda version supported by the
driver.