Fix automatic numbering of outputfiles
Before this commit, the numbering of outputfiles was achieved by storing
the current file number in a map with the saved quantity as key. This
approach leads to unexpected results when the user wants to save a
derived quantity repeatedly, e.g.:

    save(m.region(1)) // writes to m.region1000000.ovf
    save(m.region(1)) // overwrites m.region1000000.ovf

This commit solves this issue by using the quantity name as key instead
of the quantity itself.