rm sp4_madd_bench.mx3
This script is used for the benchmarks mentioned in pull request #233.

This script does not actively check anything, it takes a long time to run, and has no longer any purpose. There is no reason to have this as a 'unit test' in the test directory. Hence the file is removed.